+++ Source/Krypton Components/Krypton.Utilities/KryptonWebView2/Designers/Designers/KryptonWebView2Designer.cs
@@ -1,0 +1,36 @@
+#region BSD License
+/*
+ * 
+ *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
+ *  Modifications by Peter Wagner (aka Wagnerp), Simon Coghlan (aka Smurf-IV), Giduac & Ahmed Abdelhameed et al. 2026 - 2026. All rights reserved.
+ *  
+ */
+#endregion
+
+namespace Krypton.Utilities;
+
+#if WEBVIEW2_AVAILABLE
+internal class KryptonWebView2Designer : ControlDesigner
+{
+    #region Public Overrides
+    /// <summary>
+    /// Initializes the designer with the specified component.
+    /// </summary>
+    /// <param name="component">The IComponent to associate the designer with.</param>
+    public override void Initialize([DisallowNull] IComponent component)
+    {
+        // Let base class do standard stuff
+        base.Initialize(component);
+
+        Debug.Assert(component != null);
+
+        // The resizing handles around the control need to change depending on the
+        // value of the AutoSize and AutoSizeMode properties. When in AutoSize you
+        // do not get the resizing handles, otherwise you do.
+        AutoResizeHandles = true;
+    }
+
+    #endregion Public Overrides
+}
+#endif
+